\name{downloadcorpus}
\alias{downloadcorpus}
\title{Download Hindawi corpus zip file to "~/".}
\usage{
  downloadcorpus()
}
\description{
  Download Hindawi corpus zip file to "~/".
}
\details{
  This function is also used for updating the corpus.
}
\examples{
\dontrun{
downloadcorpus()
}
}
\author{
  Scott Chamberlain \email{myrmecocystus@gmail.com}
}

