\name{loadcorpus}
\alias{loadcorpus}
\title{Load the Hindawi corpus into R as a nested list of XML files.}
\usage{
  loadcorpus()
}
\description{
  Load the Hindawi corpus into R as a nested list of XML
  files.
}
\examples{
\dontrun{
loadcorpus()
}
}
\author{
  Scott Chamberlain \email{myrmecocystus@gmail.com}
}
\seealso{
  \code{\link{downloadcorpus}} and
  \code{\link{createrdat}}.
}
\keyword{load}

