\name{subjects}
\alias{subjects}
\title{Get the subjects areas for Hindawi journals, and the journals that
   are in each subject.}
\usage{
  subjects()
}
\description{
  Get the subjects areas for Hindawi journals, and the
  journals that are in each subject.
}
\examples{
\dontrun{
subjects()
}
}
\author{
  Scott Chamberlain \email{myrmecocystus@gmail.com}
}

